More module refactoring
Mesh has it's own module now and a lot of functionality added directly
to it's implementation which has already made the subdivide algorithm a
little bit easier to follow.

Some of the change might have some performance implication until we have
a better idea of how to reserve vectors in a sane fashion (aka, I'm not
reserving any space so now there might be some spurious reallocations).